requirements.txt
streamlit==1.55.0
openai==2.26.0
langchain==1.2.10
langchain-core==1.2.17
langchain-community==0.4.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
chromadb==1.5.4
python-dotenv==1.2.2
pandas==2.3.3
tabulate==0.10.0
tiktoken==0.12.0
